Fix synctriggers for functions
This commit adds code to make sure that synctriggers is always called
after a function app has finished deploying. Additionally, this change
also adds an empty settriggers call that is required by the front end to
notify the zip cache that changes were made. The change is current off
by default using SCMHostingConfig

+++ Kudu.Core/Infrastructure/DeploymentLockFile.cs
@@ -3,6 +3,7 @@ using System.IO;
 using System.IO.Abstractions;
 using Kudu.Contracts.SourceControl;
 using Kudu.Core.Helpers;
+using Kudu.Core.Settings;
 using Kudu.Core.SourceControl;
 using Kudu.Core.Tracing;
 
@@ -14,6 +15,8 @@ namespace Kudu.Core.Infrastructure
     public class DeploymentLockFile : LockFile
     {
         private string siteRoot = "";
+        private bool shutdownDelayed = false;
+        private readonly ITraceFactory _traceFactory;
 
         public DeploymentLockFile(string path, ITraceFactory traceFactory)
             : base(path, traceFactory)
@@ -24,6 +27,8 @@ namespace Kudu.Core.Infrastructure
                     siteRoot = PathUtilityFactory.Instance.ResolveLocalSitePath();
                 }
             });
+
+            _traceFactory = traceFactory;
         }
 
         // This is set when IRepositoryFactory is created in Ninject.
@@ -35,27 +40,33 @@ namespace Kudu.Core.Infrastructure
 
         protected override void OnLockAcquired()
         {
-
-            OperationManager.SafeExecute(() => {
-                // Create Sentinel file for DWAS to check
-                // DWAS will check for presence of this file incase a an app setting based recycle needs to be performed in middle of deployment
-                // If this is present, DWAS will postpone the recycle so that deployment goes through first
-                if (!String.IsNullOrEmpty(siteRoot))
-                {
-                    FileSystemHelpers.CreateDirectory(Path.Combine(siteRoot, @"ShutdownSentinel"));
-                    string sentinelPath = Path.Combine(siteRoot, @"ShutdownSentinel\Sentinel.txt");
-
-                    if (!FileSystemHelpers.FileExists(sentinelPath))
+            if (ScmHostingConfigurations.FunctionsSyncTriggersDelayBackground)
+            {
+                shutdownDelayed = ShutdownDelaySemaphore.GetInstance().Acquire(_traceFactory.GetTracer());
+            }
+            else
+            {
+                OperationManager.SafeExecute(() => {
+                    // Create Sentinel file for DWAS to check
+                    // DWAS will check for presence of this file incase a an app setting based recycle needs to be performed in middle of deployment
+                    // If this is present, DWAS will postpone the recycle so that deployment goes through first
+                    if (!String.IsNullOrEmpty(siteRoot))
                     {
-                        var file = FileSystemHelpers.CreateFile(sentinelPath);
-                        file.Close();
-                    }
+                        FileSystemHelpers.CreateDirectory(Path.Combine(siteRoot, @"ShutdownSentinel"));
+                        string sentinelPath = Path.Combine(siteRoot, @"ShutdownSentinel\Sentinel.txt");
 
-                    // DWAS checks if write time of this file is in the future then only postpones the recycle
-                    FileInfoBase sentinelFileInfo = FileSystemHelpers.FileInfoFromFileName(sentinelPath);
-                    sentinelFileInfo.LastWriteTimeUtc = DateTime.UtcNow.AddMinutes(20);
-                }
-            });
+                        if (!FileSystemHelpers.FileExists(sentinelPath))
+                        {
+                            var file = FileSystemHelpers.CreateFile(sentinelPath);
+                            file.Close();
+                        }
+
+                        // DWAS checks if write time of this file is in the future then only postpones the recycle
+                        FileInfoBase sentinelFileInfo = FileSystemHelpers.FileInfoFromFileName(sentinelPath);
+                        sentinelFileInfo.LastWriteTimeUtc = DateTime.UtcNow.AddMinutes(20);
+                    }
+                });
+            }
 
             IRepositoryFactory repositoryFactory = RepositoryFactory;
             if (repositoryFactory != null)
@@ -73,18 +84,25 @@ namespace Kudu.Core.Infrastructure
         {
             base.OnLockRelease();
 
-            OperationManager.SafeExecute(() => {
-                // Delete the Sentinel file to signal DWAS that deployment is complete
-                if (!String.IsNullOrEmpty(siteRoot))
-                {
-                    string sentinelPath = Path.Combine(siteRoot, @"ShutdownSentinel\Sentinel.txt");
-
-                    if (FileSystemHelpers.FileExists(sentinelPath))
+            if (ScmHostingConfigurations.FunctionsSyncTriggersDelayBackground && shutdownDelayed)
+            {
+                ShutdownDelaySemaphore.GetInstance().Release(_traceFactory.GetTracer());
+            }
+            else
+            {
+                OperationManager.SafeExecute(() => {
+                    // Delete the Sentinel file to signal DWAS that deployment is complete
+                    if (!String.IsNullOrEmpty(siteRoot))
                     {
-                        FileSystemHelpers.DeleteFile(sentinelPath);
+                        string sentinelPath = Path.Combine(siteRoot, @"ShutdownSentinel\Sentinel.txt");
+
+                        if (FileSystemHelpers.FileExists(sentinelPath))
+                        {
+                            FileSystemHelpers.DeleteFile(sentinelPath);
+                        }
                     }
-                }
-            });
+                });
+            }
         }
     }
 }
